pyproject.toml
[tool.poetry]
name = "fast-graphql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
sqlalchemy = "^1.3.19"
mysql-connector-python = "^8.0.21"
bcrypt = "^3.2.0"
pytest = "^6.0.1"
PyJWT = "^1.7.1"
graphene = "^2.1.8"
graphene-sqlalchemy = "^2.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
